STEP1: 객체미용체조 7원칙: 원시 타입(Int, Double, String...)을 두 개 이상 갖지 않도록 한다. 소포 배송비 관련 프로퍼티도 ParcelCostInfo 타입으로 통합.
