--- ParcelInvoiceMaker/ParcelInvoiceMaker/Feature/Parcel/Model/Discount.swift
@@ -1,12 +1,0 @@
-//
-//  Discount.swift
-//  ParcelInvoiceMaker
-//
-//  Created by Choi Oliver on 3/4/24.
-//
-
-import Foundation
-
-enum Discount: Int {
-    case none = 0, vip, coupon
-}
